Another incremental checkin.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1340 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/save/kernel/platform/unix/FileSystem.cpp
@@ -442,7 +442,7 @@ const char *SearchFileName(
 //}
 
 
-void SysLoadImage(char **imageBuffer, long *imageSize)
+void SysLoadImage(char **imageBuffer, size_t *imageSize)
 /*******************************************************************/
 /* Function : Load the image into storage                          */
 /*******************************************************************/
